feat: added restart-ms to candevice module
There was a placeholder "RestartMs" field in candevice.Info but was
unused.
This change implements a SetRestartMs(time.Duration) function, and the
value can be read back from the candevice.Info struct.

The type of the value that the caller interacts with is time.Duration,
which is convenient in go. The function name is left "RestartMs" as a
reminder of the effective granularity of the duration, and to not
interfere with the "Restart" command that might be implemented here in
the future.